+++ package.json
@@ -23,7 +23,7 @@
     "react-bootstrap-tabs": "^2.0.0",
     "react-dom": "^18.2.0",
     "react-firebase-hooks": "^5.1.1",
-    "react-redux": "^8.0.5",
+    "react-redux": "^8.1.1",
     "react-router-dom": "^6.8.1",
     "react-scripts": "5.0.1",
     "react-scroll": "^1.8.9",
