+++ src/App.js
@@ -23,6 +23,7 @@ import ScrollToTop from './Pages/scrollToTop';
 import NavigatorMenu from './Components/Navigation/NavigatorMenu';
 import MilestoneMap from './Components/Milestones/MilestoneMap';
 import Internships from './Components/Internships/Internships';
+import Assessment from './Components/Assessment/Assessment';
 import Workshops from './Components/Workshops/Workshops';
 import JobFairs from './Components/JobFairs/JobFairs';
 import JobBoards from './Components/JobBoards/JobBoards';
@@ -91,9 +92,12 @@ function App() {
             </>
           )}
         />
+        {user != undefined && 
+        (
+        <>
         <Route path="/home" element={<NavigatorMenu />}>
           <Route path="roadmap" element={<MilestoneMap />} />
-          <Route path="assessment" element={<JobseekerData />} />
+          <Route path="assessment" element={<Assessment />} />
           <Route path="onlineprofiles" element={<OnlineProfiles />} />
           <Route path="training" element={<TrainingPrograms />} />
           <Route path="internships" element={<Internships />} />
@@ -103,11 +107,13 @@ function App() {
           <Route path="resources" element={<Resources />} />
           <Route path="hiredinfo" element={<HiredInfo />} />
         </Route>
+        <Route path="/onboard" element={<JobseekerData />} />
+      </>)
+}
         <Route path="/login" element={<Login />} />
         <Route path="/dashboard/navigator" element={<NavigatorDashboard />} />
         <Route path="/reset" element={<Reset />} />
         <Route path="/landing" element={<Landing />} />
-        <Route path="/onboard" element={<JobseekerData useremail="solia@goodpl.us" username="Solia Nasser" />} />
         <Route path="/adminview" element={<AdminView />} />
 
         <Route
